Integration with search engine
We've still a lengthy to-do list. Need to include all results in our
list of sorted documents (those with a score of 0), plus implementing
other precision measurements and generating random results. Also, the
search "Pirates of the Caribbean: Dead Man's Chest" is returning no
results for no immediately discernable reason (I have adjusted the query
formatting function to handle punctuation, so it's not that, as far as I
know)